made some changes in implementation

+++ src/Wallet.java
@@ -1,7 +1,7 @@
  interface Wallet{
 
     public long addMoney ( long money);
-    public long removeMoney ( long money);
+    public int removeMoney (long money);
     public long getBalance ();
 
 }
